Update server_updateObject.sqf
You should disable this. You know about the bug that sometimes due to lag, you buy a vehicle from a trader and it disappears ? And then on restart the vehicle is back at the location it previously spawned ?
This little line causes that. I've disabled that since 1.0.4.2 and i've never had that problem again :)